requirements.txt
Django==5.0.6
psycopg2-binary==2.9.9
django-environ==0.11.2
django-htmx==1.17.3
django-compressor==4.4
whitenoise==6.6.0
django-htmlmin==0.11.0